Start refactor pushbuttons for testing
