+++ internal/pushbuttons/pushbuttons_test.go
@@ -1,9 +1,6 @@
-package filechecks
+package pushbuttons
 
-import (
-	"testing"
-)
-
+/*
 func TestGetFileInfos(t *testing.T) {
 	tests := []struct {
 		name          string
@@ -42,3 +39,4 @@ func TestGetFileInfos(t *testing.T) {
 		})
 	}
 }
+*/
